test(canon): split matrix evidence coverage tests
