+++ tests/tooling/release-preflight-matrix-evidence.test.ts
@@ -1,13 +1,16 @@
 import assert from "node:assert/strict";
-import { createHash } from "node:crypto";
 import { test } from "node:test";
 
 import {
   assertRealProjectMatrixReleaseArtifacts,
   requireRealProjectMatrixRun,
-  requiredRealProjectMatrixShardCount,
 } from "../../tools/github/release-preflight-matrix-evidence.mjs";
-import { releaseSha, successfulReleaseRun } from "./support/release-preflight.ts";
+import {
+  mutateDivergence,
+  realProjectArtifacts,
+  shardEntries,
+} from "./_helpers/release-preflight-matrix-evidence-fixture.ts";
+import { successfulReleaseRun } from "./support/release-preflight.ts";
 
 test("release preflight validates every real-project shard artifact", async () => {
   const run = successfulReleaseRun("Real Project Matrix", 500);
@@ -113,137 +116,6 @@ test("release preflight rejects record-only and non-zero typecheck parity artifa
   }
 });
 
-test("release preflight requires same-corpus coverage, mutation oracle, and preparation linkage", async () => {
-  for (const [label, mutate, message] of [
-    [
-      "coverage mismatch",
-      (entries: Record<string, string>) =>
-        mutateDivergence(
-          entries,
-          (artifact) => (artifact.baseline.coverage.baselineVueFilesSha256 = "0".repeat(64)),
-        ),
-      /same non-empty authored Vue corpus/,
-    ],
-    [
-      "empty shared coverage",
-      (entries: Record<string, string>) =>
-        mutateDivergence(entries, (artifact) => {
-          artifact.baseline.coverage.sharedVueFileCount = 0;
-          artifact.baseline.coverage.vizeVueFileCount = 0;
-          artifact.baseline.coverage.baselineVueFileCount = 0;
-        }),
-      /same non-empty authored Vue corpus/,
-    ],
-    [
-      "missing mutation",
-      (entries: Record<string, string>) =>
-        mutateDivergence(entries, (artifact) => delete artifact.mutationOracle),
-      /seeded mutation oracle/,
-    ],
-    [
-      "forged broken mutation state",
-      (entries: Record<string, string>) =>
-        mutateDivergence(
-          entries,
-          (artifact) => (artifact.mutationOracle.states[1].sharedCount = 0),
-        ),
-      /seeded mutation oracle/,
-    ],
-    [
-      "unrestored repaired mutation state",
-      (entries: Record<string, string>) =>
-        mutateDivergence(
-          entries,
-          (artifact) => (artifact.mutationOracle.states[2].sourceSha256 = "7".repeat(64)),
-        ),
-      /seeded mutation oracle/,
-    ],
-    [
-      "unchanged broken mutation source",
-      (entries: Record<string, string>) =>
-        mutateDivergence(
-          entries,
-          (artifact) =>
-            (artifact.mutationOracle.states[1].sourceSha256 =
-              artifact.mutationOracle.states[0].sourceSha256),
-        ),
-      /seeded mutation oracle/,
-    ],
-    [
-      "unhashed mutation source",
-      (entries: Record<string, string>) =>
-        mutateDivergence(
-          entries,
-          (artifact) => (artifact.mutationOracle.states[1].sourceSha256 = "not-a-digest"),
-        ),
-      /seeded mutation oracle/,
-    ],
-    [
-      "diagnostic-bearing clean mutation state",
-      (entries: Record<string, string>) =>
-        mutateDivergence(
-          entries,
-          (artifact) => (artifact.mutationOracle.states[0].sharedCount = 1),
-        ),
-      /seeded mutation oracle/,
-    ],
-    [
-      "diagnostic-bearing repaired mutation state",
-      (entries: Record<string, string>) =>
-        mutateDivergence(
-          entries,
-          (artifact) => (artifact.mutationOracle.states[2].sharedCount = 1),
-        ),
-      /seeded mutation oracle/,
-    ],
-    [
-      "failed surface verdict",
-      (entries: Record<string, string>) => {
-        entries["surface-verdict.json"] = json({ status: "failure" });
-      },
-      /surface verdict is failure/,
-    ],
-    [
-      "wrong shard count",
-      (entries: Record<string, string>) => {
-        const summary = JSON.parse(entries["summary.json"]);
-        summary.command.shardCount = requiredRealProjectMatrixShardCount - 1;
-        entries["summary.json"] = json(summary);
-      },
-      /not exact release evidence/,
-    ],
-    [
-      "missing dependency artifact",
-      (entries: Record<string, string>) => delete entries["fixture-typecheck-dependencies.json"],
-      /typecheck dependency artifact/,
-    ],
-    [
-      "unlinked dependency artifact",
-      (entries: Record<string, string>) =>
-        mutateDivergence(
-          entries,
-          (artifact) => (artifact.preparation.payloadSha256 = "0".repeat(64)),
-        ),
-      /missing dependency preparation linkage/,
-    ],
-  ] as const) {
-    const run = successfulReleaseRun("Real Project Matrix", 520);
-    await assert.rejects(
-      assertRealProjectMatrixReleaseArtifacts({
-        run,
-        artifacts: realProjectArtifacts(run),
-        readArtifactEntries: async (artifact) => {
-          const entries = shardEntries(Number(artifact.name.split("-").pop()));
-          if (artifact.name === "real-project-matrix-0") mutate(entries);
-          return entries;
-        },
-      }),
-      message,
-      label,
-    );
-  }
-});
-
 test("release preflight requires a selected Real Project Matrix run", () => {
   const run = successfulReleaseRun("Real Project Matrix", 530);
   assert.equal(requireRealProjectMatrixRun(new Map([["Real Project Matrix", run]])), run);
@@ -254,141 +126,3 @@ test("release preflight requires a selected Real Project Matrix run", () => {
     );
   }
 });
-
-function realProjectArtifacts(run: ReturnType<typeof successfulReleaseRun>) {
-  return Array.from({ length: requiredRealProjectMatrixShardCount }, (_, shard) => ({
-    id: 1_000 + shard,
-    name: `real-project-matrix-${shard}`,
-    expired: false,
-    archive_download_url: `https://example.test/artifacts/${shard}.zip`,
-    workflow_run: {
-      id: run.id,
-      head_branch: run.head_branch,
-      head_sha: run.head_sha,
-    },
-  }));
-}
-
-function shardEntries(shard: number): Record<string, string> {
-  const dependency = {
-    schema: "vize.fixtureTypecheckDependencyInstall",
-    version: 2,
-    project: "fixture",
-    revision: "b".repeat(40),
-    evidence: { commitSha: releaseSha },
-    packageManager: { name: "pnpm", version: "10.0.0" },
-    lockfile: { path: "pnpm-lock.yaml", sizeBytes: 18, sha256: "1".repeat(64) },
-    install: {
-      command: ["pnpm", "install", "--frozen-lockfile", "--ignore-scripts", "--prefer-offline"],
-      durationMs: 1,
-      exitCode: 0,
-      stdoutSha256: "2".repeat(64),
-      stderrSha256: "3".repeat(64),
-    },
-    baselinePrepare: null,
-  };
-  const dependencyText = json(dependency);
-  return {
-    "selected-fixtures.txt": "tests/_fixtures/_git/fixture\n",
-    "summary.json": json({
-      schema: "vize.fixtureToolMatrixReport",
-      version: 3,
-      evidence: { commitSha: releaseSha },
-      command: { shardIndex: shard, shardCount: requiredRealProjectMatrixShardCount },
-    }),
-    "surface-verdict.json": json({ status: "success" }),
-    "fixture-typecheck-dependencies.json": dependencyText,
-    "fixture-typecheck-divergence.json": json({
-      schema: "vize.fixtureTypecheckDivergenceRun",
-      version: 4,
-      project: "fixture",
-      revision: "b".repeat(40),
-      evidence: { commitSha: releaseSha },
-      enforcement: { budgetMode: "enforce" },
-      preparation: {
-        schema: "vize.fixtureTypecheckPreparationEvidence",
-        version: 1,
-        payloadSha256: sha256(dependencyText),
-      },
-      baseline: {
-        coverage: {
-          verdict: "usable",
-          vizeVueFileCount: 1,
-          baselineVueFileCount: 1,
-          sharedVueFileCount: 1,
-          vizeVueFilesSha256: "4".repeat(64),
-          baselineVueFilesSha256: "4".repeat(64),
-          missingVueFiles: [],
-          unexpectedVueFiles: [],
-        },
-      },
-      mutationOracle: {
-        schema: "vize.fixtureTypecheckSeededMutationOracle",
-        version: 1,
-        verdict: "passed",
-        passed: true,
-        file: "src/App.vue",
-        span: { line: 3, column: 1 },
-        cleanExpectedDiagnosticPresent: false,
-        expectedDiagnosticMatched: true,
-        repairedExpectedDiagnosticPresent: false,
-        states: [
-          mutationState("clean", "5".repeat(64), 0, 0, 0),
-          mutationState("broken", "6".repeat(64), 1, 0, 0),
-          mutationState("repaired", "5".repeat(64), 0, 0, 0),
-        ],
-      },
-      budget: { passed: true, verdict: "passed" },
-      divergence: {
-        summary: {
-          falsePositiveCount: 0,
-          falseNegativeCount: 0,
-        },
-      },
-    }),
-  };
-}
-
-function mutationState(
-  name: string,
-  sourceSha256: string,
-  sharedCount: number,
-  falsePositiveCount: number,
-  falseNegativeCount: number,
-) {
-  const summary = {
-    vizeDiagnosticCount: sharedCount,
-    baselineDiagnosticCount: sharedCount,
-    sharedCount,
-    messageMismatchCount: 0,
-    documentedDifferenceCount: 0,
-    falsePositiveCount,
-    falseNegativeCount,
-  };
-  return {
-    name,
-    sourceSha256,
-    ...summary,
-    observed: summary,
-    vize: mutationRun("vize", falsePositiveCount > 0 ? 1 : 0),
-    baseline: mutationRun("vue-tsc", falseNegativeCount > 0 ? 2 : 0),
-  };
-}
-
-function mutationRun(command: string, exitCode: number) {
-  return { command, exitCode, stdoutSha256: "8".repeat(64), stderrSha256: "9".repeat(64) };
-}
-
-function mutateDivergence(entries: Record<string, string>, mutate: (artifact: any) => void) {
-  const artifact = JSON.parse(entries["fixture-typecheck-divergence.json"]);
-  mutate(artifact);
-  entries["fixture-typecheck-divergence.json"] = json(artifact);
-}
-
-function json(value: unknown) {
-  return `${JSON.stringify(value, null, 2)}\n`;
-}
-
-function sha256(value: string) {
-  return createHash("sha256").update(value).digest("hex");
-}
